created parsReques.js and test  for that
